fix(#296): server-render copy em /classificacao e /calculadora
Antes: ambas as páginas eram client components puros e serviam ~0 e 61
palavras ao Googlebot — schema do PR #295 sem conteúdo correspondente.

Mudanças:
- /classificacao/page.tsx → Server Component com hero, 5 H2 explicativos,
  tabela de estrutura cClassTrib, FAQ (6 Q&A) e fundamentação legal.
  Widget interativo isolado em Classifier.tsx (client).
- /calculadora/page.tsx → Server Component com hero, 4 H2, 2 tabelas
  (alíquotas, regimes), FAQ (6 Q&A) e fundamentação legal. Widget em
  Calculator.tsx.
- AppShell.tsx: adicionar /classificacao em PUBLIC_ROUTES (estava como
  rota protegida, causando loading-state + redirect ao login no SSR).
- Novo componente reutilizável FundamentacaoLegal.tsx (table semântica).

Resultado pós-build (SSG):
- /classificacao: 0 → 1.025 palavras, 0 → 14 headings, 0 → 2 tabelas
- /calculadora: 61 → 776 palavras, 0 → 13 headings, 0 → 3 tabelas
